+++ noodles-bam/src/record/builder.rs
@@ -15,7 +15,7 @@ pub struct Builder {
     mapq: Option<sam::record::MappingQuality>,
     flag: sam::record::Flags,
     next_ref_id: Option<ReferenceSequenceId>,
-    next_pos: Option<sam::record::Position>,
+    next_pos: Option<Position>,
     tlen: i32,
     read_name: Option<sam::record::ReadName>,
     cigar: sam::record::Cigar,
@@ -140,16 +140,16 @@ impl Builder {
     ///
     /// ```
     /// use noodles_bam as bam;
-    /// use noodles_sam as sam;
+    /// use noodles_core::Position;
     ///
     /// let record = bam::Record::builder()
-    ///     .set_mate_position(sam::record::Position::try_from(13)?)
+    ///     .set_mate_position(Position::try_from(13)?)
     ///     .build()?;
     ///
-    /// assert_eq!(record.mate_position().map(i32::from), Some(13));
+    /// assert_eq!(record.mate_position(), Position::new(13));
     /// # Ok::<(), Box<dyn std::error::Error>>(())
     /// ```
-    pub fn set_mate_position(mut self, mate_position: sam::record::Position) -> Self {
+    pub fn set_mate_position(mut self, mate_position: Position) -> Self {
         self.next_pos = Some(mate_position);
         self
     }